Fix VSTestTask to use DOTNET_HOST_PATH and revert BuildEnvironmentHelper workaround
VSTestTask: Use DOTNET_HOST_PATH environment variable (set by the SDK) to
resolve the dotnet host path, with PATH search as fallback. This is the
proper fix for the 'dotnet not found' error that occurs when MSBuild runs
via the apphost and Environment.ProcessPath changes to the SDK directory.

BuildEnvironmentHelper: Revert the IsRunningInMSBuildExe guard in
TryFromMSBuildAppHost and the context-aware logic in
TryFromAppContextBaseDirectory. These were workarounds for the same issue
and are no longer needed now that VSTestTask resolves dotnet properly.
