+++ src/msbuild/src/Shared/BuildEnvironmentHelper.cs
@@ -208,17 +208,6 @@ namespace Microsoft.Build.Shared
                 return environment;
             }
 
-            // Only use the app host if we're actually running as the app host.
-            // When running via dotnet (e.g. dotnet msbuild), prefer MSBuild.dll so that
-            // worker nodes inherit the dotnet process context — otherwise
-            // Environment.ProcessPath changes to the SDK directory where "dotnet" isn't present,
-            // breaking Process.Start("dotnet") resolution in tasks like VSTestTask.
-            bool runningAsAppHost = IsRunningInMSBuildExe(EnvironmentUtilities.ProcessPath);
-            if (!runningAsAppHost)
-            {
-                return null;
-            }
-
             // We're not in VS, check for MSBuild.exe / dll to consider this a standalone environment.
             string msBuildPath = null;
             if (FileSystems.Default.FileExists(msBuildExecutableCandidate))
@@ -337,18 +326,10 @@ namespace Microsoft.Build.Shared
                 return null;
             }
 
-            // When running as the MSBuild app host (native executable), prefer the app host.
-            // When running via dotnet (e.g. dotnet msbuild, dotnet test), prefer MSBuild.dll
-            // so that worker nodes inherit the dotnet process context — otherwise
-            // Environment.ProcessPath changes to the SDK directory where "dotnet" isn't present,
-            // breaking Process.Start("dotnet") resolution in tasks like VSTestTask.
-            bool runningAsAppHost = IsRunningInMSBuildExe(EnvironmentUtilities.ProcessPath);
-
-            string primary = Path.Combine(appContextBaseDirectory, runningAsAppHost ? Constants.MSBuildExecutableName : Constants.MSBuildAssemblyName);
-            string fallback = Path.Combine(appContextBaseDirectory, runningAsAppHost ? Constants.MSBuildAssemblyName : Constants.MSBuildExecutableName);
-
-            return TryFromStandaloneMSBuildExe(primary)
-                ?? TryFromStandaloneMSBuildExe(fallback);
+            // Prioritize MSBuild[.exe] over MSBuild.dll
+            return TryFromStandaloneMSBuildExe(Path.Combine(appContextBaseDirectory, Constants.MSBuildExecutableName))
+                // Fall back to MSBuild.dll
+                ?? TryFromStandaloneMSBuildExe(Path.Combine(appContextBaseDirectory, Constants.MSBuildAssemblyName));
         }
 
         private static BuildEnvironment TryFromStandaloneMSBuildExe(string msBuildExePath)
